Move UserConnectionRepository tests to an abstract base class
Allows easier testing of own implementations
