+++ spring-social-core/src/main/java/org/springframework/social/connect/jdbc/JdbcUsersConnectionRepository.java
@@ -60,16 +60,11 @@ public class JdbcUsersConnectionRepository implements UsersConnectionRepository 
 		this.textEncryptor = textEncryptor;
 	}
 
-	/**
-	 * The command to execute to create a new local user profile in the event no user id could be mapped to a connection.
-	 * Allows for implicitly creating a user profile from connection data during a provider sign-in attempt.
-	 * Defaults to null, indicating explicit sign-up will be required to complete the provider sign-in attempt.
-	 * @see #findUserIdsWithConnection(Connection)
-	 */
+	@Override
 	public void setConnectionSignUp(ConnectionSignUp connectionSignUp) {
 		this.connectionSignUp = connectionSignUp;
 	}
-	
+
 	/**
 	 * Sets a table name prefix. This will be prefixed to all the table names before queries are executed. Defaults to "".
 	 * This is can be used to qualify the table name with a schema or to distinguish Spring Social tables from other application tables.
