+++ spring-social-core/src/main/java/org/springframework/social/connect/mem/InMemoryUsersConnectionRepository.java
@@ -50,12 +50,7 @@ public class InMemoryUsersConnectionRepository implements UsersConnectionReposit
 		this.connectionRepositories= new HashMap<String, ConnectionRepository>(); 
 	}
 	
-	/**
-	 * The command to execute to create a new local user profile in the event no user id could be mapped to a connection.
-	 * Allows for implicitly creating a user profile from connection data during a provider sign-in attempt.
-	 * Defaults to null, indicating explicit sign-up will be required to complete the provider sign-in attempt.
-	 * @see #findUserIdsWithConnection(Connection)
-	 */
+	@Override
 	public void setConnectionSignUp(ConnectionSignUp connectionSignUp) {
 		this.connectionSignUp = connectionSignUp;
 	}
